+++ autode/transition_states/ts_guess.py
@@ -4,9 +4,10 @@ from autode.transition_states.templates import get_ts_templates
 from autode.transition_states.templates import template_matches
 from autode.calculation import Calculation
 from autode.config import Config
+from autode.values import Distance
 from autode.exceptions import CalculationException
 from autode.log import logger
-from autode.values import Distance
+from autode.utils import work_in
 from autode.methods import get_lmethod, get_hmethod
 from autode.mol_graphs import (get_mapping_ts_template,
                                get_truncated_active_mol_graph)
@@ -42,22 +43,22 @@ def get_template_ts_guess(reactant:   'autode.species.ReactantComplex',
                           product:    'autode.species.ProductComplex',
                           bond_rearr: 'autode.bond_rearrangement.BondRearrangement',
                           name:       str,
-                          method:     'autode.wrappers.base.ElectronicStructureMethod',
-                          dist_thresh: Distance = Distance(4.0, units='Å')):
-    """Get a transition state guess object by searching though the stored TS
+                          method:     'autode.wrappers.base.ElectronicStructureMethod'):
+    """
+    Get a transition state guess object by searching though the stored TS
     templates
 
+    ---------------------------------------------------------------------------
     Arguments:
         reactant (autode.complex.ReactantComplex):
-        bond_rearr (autode.bond_rearrangement.BondRearrangement):
-        product (autode.complex.ProductComplex):
-        method (autode.wrappers.base.ElectronicStructureMethod):
-        name (str):
 
-    Keyword Arguments:
-        dist_thresh (float): distance above which a constrained optimisation
-                             probably won't work due to the initial geometry
-                             being too far away from the ideal
+        bond_rearr (autode.bond_rearrangement.BondRearrangement):
+
+        product (autode.complex.ProductComplex):
+
+        method (autode.wrappers.base.ElectronicStructureMethod):
+
+        name (str):
 
     Returns:
         (autode.transition_states.ts_guess.TSguess):
@@ -96,20 +97,14 @@ def get_template_ts_guess(reactant:   'autode.species.ReactantComplex',
             continue
 
         logger.info('Found a TS guess from a template')
-        if any([abs(reactant.distance(*bond) - active_bonds_and_dists_ts[bond])
-                > dist_thresh for bond in bond_rearr.all]):
-            logger.info(f'TS template has => 1 active bond distance larger '
-                        f'than {dist_thresh}. Passing')
-            continue
-
         ts_guess = TSguess(atoms=reactant.atoms,
                            reactant=reactant,
                            product=product,
                            bond_rearr=bond_rearr)
+        ts_guess.constraints.distance = active_bonds_and_dists_ts
 
         try:
             ts_guess.run_constrained_opt(name=name,
-                                         distance_consts=active_bonds_and_dists_ts,
                                          method=method,
                                          keywords=method.keywords.opt)
             return ts_guess
@@ -124,19 +119,56 @@ def get_template_ts_guess(reactant:   'autode.species.ReactantComplex',
 class TSguess(TSbase):
     """Transition state guess"""
 
+    @work_in('scan_to_template')
+    def _lmethod_scan_to_point(self):
+        """
+        Run a set of constrained low-level optimisations from the current
+        distances to the final set of constraints using a linear path with
+        small distance increments.
+
+        Raises:
+            (autode.exceptions.CalculationException):
+        """
+        l_method = get_lmethod()
+
+        # Final and current sets of constraints
+        f_consts = self.constraints.distance
+        c_consts = {bond: self.distance(*bond) for bond in f_consts.keys()}
+
+        # Number of steps to use is 0.1 Å in the maximum distance delta
+        n_steps = ((max(abs(f_consts[bond] - c_dist)
+                        for bond, c_dist in c_consts.items()))
+                   / Distance(0.1, units='ang'))
+
+        for i in range(1, n_steps+1):
+
+            consts = {}
+            for bond in c_consts.keys():
+                consts[bond] = i * (f_consts[bond] - c_consts[bond]) / n_steps
+
+            opt = Calculation(name=f'{self.name}_const_opt_ll_{i}',
+                              molecule=self,
+                              method=l_method,
+                              keywords=l_method.keywords.low_opt,
+                              n_cores=Config.n_cores,
+                              distance_constraints=consts)
+
+            # Can raise CalculationException
+            self.optimise(calc=opt)
+
+        return None
+
     def run_constrained_opt(self,
-                            name:            str,
-                            distance_consts: dict,
-                            method:          Optional['autode.wrappers.base.ElectronicStructureMethod'] = None,
-                            keywords:        Optional['autode.wrappers.keywords.Keywords'] = None):
+                            name:     str,
+                            method:   Optional['autode.wrappers.base.ElectronicStructureMethod'] = None,
+                            keywords: Optional['autode.wrappers.keywords.Keywords'] = None):
         """Get a TS guess from a constrained optimisation with the active atoms
         fixed at values defined in distance_consts
 
         Arguments:
-            keywords (autode.wrappers.keywords.Keywords):
             name (str):
-            distance_consts (dict): Distance constraints keyed with a tuple of
-                                    atom indexes and value of the distance
+
+            keywords (autode.wrappers.keywords.Keywords):
 
         Keyword Arguments:
 
@@ -149,31 +181,11 @@ class TSguess(TSbase):
         Raises:
             (autode.exceptions.CalculationException):
         """
-        logger.info('Getting TS guess from constrained optimisation')
+        logger.info('Running constrained optimisation on TS guess geometry')
 
-        # Run a low level constrained optimisation first to prevent slow
-        # high-level optimisation for a TS that is far from the current
-        # geometry. If XTB then use a small-ish force constant
-        l_method = get_lmethod()
-        if l_method.name == 'xtb':
-            l_method.force_constant = 0.1
+        self._lmethod_scan_to_point()
 
-        ll_const_opt = Calculation(name=f'{name}_constrained_opt_ll',
-                                   molecule=self,
-                                   method=l_method,
-                                   keywords=l_method.keywords.low_opt,
-                                   n_cores=Config.n_cores,
-                                   distance_constraints=distance_consts)
-
-        # Try and set the atoms, but continue if they're not found as hopefully
-        # the high-level method will be fine(?)
-        try:
-            self.optimise(calc=ll_const_opt)
-
-        except CalculationException:
-            logger.error('Failed to optimise with the low level method')
-
-        # Default to high-level optimisations
+        # Default to high-level regular optimisations
         if method is None:
             method = get_hmethod()
         if keywords is None:
@@ -183,10 +195,8 @@ class TSguess(TSbase):
                                    molecule=self,
                                    method=method,
                                    keywords=keywords,
-                                   n_cores=Config.n_cores,
-                                   distance_constraints=distance_consts)
+                                   n_cores=Config.n_cores)
 
         self.optimise(calc=hl_const_opt)
-
-        self.name = f'ts_guess_{name}'
-        return
+        self.constraints.distance.clear()
+        return None